requirements.txt
Menu
pandas
opencv-python
pillow
openpyxl